requirements.txt
Django==5.2
djangorestframework==3.15.2
drf-yasg==1.21.15
django-cors-headers==4.3.0
mysqlclient==2.2.4
python-decouple==3.8
paho-mqtt==1.6.1
pandas==3.0.2
scikit-learn==1.8.0